+++ presentation/src/main/java/com/threegap/bitnagil/presentation/writeroutine/WriteRoutineScreen.kt
@@ -41,8 +41,8 @@ import com.threegap.bitnagil.presentation.writeroutine.model.Day
 import com.threegap.bitnagil.presentation.writeroutine.model.RepeatType
 import com.threegap.bitnagil.presentation.writeroutine.model.Time
 import com.threegap.bitnagil.presentation.writeroutine.model.WriteRoutineType
-import com.threegap.bitnagil.presentation.writeroutine.model.mvi.WriteRoutineSideEffect
-import com.threegap.bitnagil.presentation.writeroutine.model.mvi.WriteRoutineState
+import com.threegap.bitnagil.presentation.writeroutine.contract.WriteRoutineSideEffect
+import com.threegap.bitnagil.presentation.writeroutine.contract.WriteRoutineState
 import org.orbitmvi.orbit.compose.collectAsState
 import org.orbitmvi.orbit.compose.collectSideEffect
 
@@ -359,7 +359,7 @@ private fun getSubRoutinePlaceHolder(index: Int): String {
 fun WriteRoutineScreenPreview() {
     BitnagilTheme {
         WriteRoutineScreen(
-            state = WriteRoutineState.Init.copy(periodUiExpanded = true, startTimeUiExpanded = true),
+            state = WriteRoutineState.INIT.copy(periodUiExpanded = true, startTimeUiExpanded = true),
             setRoutineName = {},
             setSubRoutineName = { _, _ -> },
             selectRepeatTime = {},
