+++ presentation/src/main/java/com/threegap/bitnagil/presentation/writeroutine/contract/WriteRoutineSideEffect.kt
@@ -1,0 +1,6 @@
+package com.threegap.bitnagil.presentation.writeroutine.contract
+
+sealed interface WriteRoutineSideEffect {
+    data object MoveToPreviousScreen : WriteRoutineSideEffect
+    data class ShowToast(val message: String) : WriteRoutineSideEffect
+}
