--- presentation/src/main/java/com/threegap/bitnagil/presentation/writeroutine/model/mvi/WriteRoutineSideEffect.kt
@@ -1,6 +1,0 @@
-package com.threegap.bitnagil.presentation.writeroutine.model.mvi
-
-sealed class WriteRoutineSideEffect {
-    data object MoveToPreviousScreen : WriteRoutineSideEffect()
-    data class ShowToast(val message: String) : WriteRoutineSideEffect()
-}
